Remove unneeded casts

+++ tests/tt_metal/tt_metal/deployment/dram/test_dram.cpp
@@ -130,7 +130,7 @@ static void accumulate_pattern_timing_summary(DramPatternTimingSummary& dst, con
 
         dst.ticks[r.pattern_id] += r.job_total_ticks;
         dst.jobs[r.pattern_id] += 1u;
-        dst.bytes[r.pattern_id] += static_cast<uint64_t>(r.words_checked) * sizeof(uint32_t);
+        dst.bytes[r.pattern_id] += (uint64_t)r.words_checked * sizeof(uint32_t);
     }
 }
 
@@ -154,7 +154,7 @@ static uint32_t get_dram_test_loops_from_env(uint32_t default_loops = 1u) {
         return default_loops;
     }
 
-    return static_cast<uint32_t>(value);
+    return value;
 }
 
 static bool pattern_timing_enabled() {
@@ -189,7 +189,7 @@ static void log_pattern_timing_summary(
     double raw_total_ms = 0.0;
 
     for (uint32_t pattern_id = 0; pattern_id < DramPatternTimingSummary::kMaxPatternId; ++pattern_id) {
-        raw_total_ms += static_cast<double>(summary.ticks[pattern_id]) * 1000.0 / kBlackholeClockHz;
+        raw_total_ms += summary.ticks[pattern_id] * 1000.0 / kBlackholeClockHz;
     }
 
     const double scale = (raw_total_ms > 0.0 && test_wall_ms > 0.0) ? (test_wall_ms / raw_total_ms) : 1.0;
@@ -210,7 +210,7 @@ static void log_pattern_timing_summary(
             continue;
         }
 
-        const double raw_pattern_ms = static_cast<double>(summary.ticks[pattern_id]) * 1000.0 / kBlackholeClockHz;
+        const double raw_pattern_ms = summary.ticks[pattern_id] * 1000.0 / kBlackholeClockHz;
         const double scaled_pattern_ms = raw_pattern_ms * scale;
 
         log_info(
@@ -280,11 +280,11 @@ static std::string format_error_pct(double pct) {
 static void accumulate_bank_summary(DramBankSummary& dst, const DramBaseResult& result) {
     dst.pass &= (result.failures == 0u);
 
-    dst.checked_bytes += static_cast<uint64_t>(result.words_checked) * sizeof(uint32_t);
+    dst.checked_bytes += (uint64_t)result.words_checked * sizeof(uint32_t);
 
-    dst.suspected_write_error_bytes += static_cast<uint64_t>(result.suspected_write_failures) * sizeof(uint32_t);
+    dst.suspected_write_error_bytes += (uint64_t)result.suspected_write_failures * sizeof(uint32_t);
 
-    dst.suspected_read_error_bytes += static_cast<uint64_t>(result.suspected_read_failures) * sizeof(uint32_t);
+    dst.suspected_read_error_bytes += (uint64_t)result.suspected_read_failures * sizeof(uint32_t);
 }
 
 static DramChipSummary make_chip_bank_summary(
@@ -695,7 +695,7 @@ static uint32_t get_dram_max_chips_from_env(uint32_t available_chips) {
     TT_FATAL(value > 0, "DRAM_TEST_MAX_CHIPS must be > 0");
     TT_FATAL(value <= available_chips, "DRAM_TEST_MAX_CHIPS={} exceeds available chips={}", value, available_chips);
 
-    return static_cast<uint32_t>(value);
+    return value;
 }
 
 static void accumulate_galaxy_summary(DramGalaxySummary& dst, const DramMultiInstanceSummary& run, uint64_t jobs) {
@@ -942,7 +942,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentOptimalWorkersAllDramBanks)
         accumulate_galaxy_summary(chip_summary, run, jobs.size());
         accumulate_pattern_timing_summary(chip_pattern_timing, run);
 
-        chip_summary.chips.push_back(make_chip_bank_summary(device, static_cast<uint32_t>(assignments.size()), run));
+        chip_summary.chips.push_back(make_chip_bank_summary(device, assignments.size(), run));
 
         const auto& s = run.summary;
 
@@ -1078,7 +1078,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentAllWorkersSingleDramSequent
 
     log_info(tt::LogTest, "Persistent all-workers single-DRAM sequential sweep running on {} chip(s)", devices_.size());
 
-    const uint32_t chips_to_test = get_dram_max_chips_from_env(static_cast<uint32_t>(devices_.size()));
+    const uint32_t chips_to_test = get_dram_max_chips_from_env(devices_.size());
 
     const bool parallel_chips = true;
 
@@ -1135,8 +1135,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentAllWorkersSingleDramSequent
             total_bytes_per_controller,
             chunk_bytes);
 
-        const uint64_t bytes_per_core_base =
-            (static_cast<uint64_t>(total_bytes_per_controller) / worker_cores.size()) & ~0xFFFULL;
+        const uint64_t bytes_per_core_base = (total_bytes_per_controller / worker_cores.size()) & ~0xFFFULL;
 
         TT_FATAL(
             bytes_per_core_base >= chunk_bytes,
@@ -1147,7 +1146,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentAllWorkersSingleDramSequent
         TT_FATAL(bytes_per_core_base <= std::numeric_limits<uint32_t>::max(), "bytes_per_core_base must fit uint32_t");
 
         const uint64_t covered_bytes = bytes_per_core_base * worker_cores.size();
-        const uint64_t remainder_bytes = static_cast<uint64_t>(total_bytes_per_controller) - covered_bytes;
+        const uint64_t remainder_bytes = total_bytes_per_controller - covered_bytes;
 
         TT_FATAL((remainder_bytes & 0xFFFULL) == 0ULL, "remainder_bytes must stay 4KB aligned");
 
@@ -1199,9 +1198,9 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentAllWorkersSingleDramSequent
 
                             job.job_id = job_id++;
                             job.bank_id = bank_id;
-                            job.bank_offset_lo = static_cast<uint32_t>(bank_offset & 0xFFFFFFFFull);
-                            job.bank_offset_hi = static_cast<uint32_t>((bank_offset >> 32) & 0xFFFFFFFFull);
-                            job.total_bytes = static_cast<uint32_t>(bytes_this_core);
+                            job.bank_offset_lo = bank_offset & 0xFFFFFFFFull;
+                            job.bank_offset_hi = (bank_offset >> 32) & 0xFFFFFFFFull;
+                            job.total_bytes = bytes_this_core;
                             job.chunk_bytes = chunk_bytes;
                             job.pattern_id = pattern_id;
                             job.seed = seed;
@@ -1224,7 +1223,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentAllWorkersSingleDramSequent
                 seed += advance_seed;
             }
 
-            const uint64_t bank_jobs = static_cast<uint64_t>(job_id - 1u);
+            const uint64_t bank_jobs = job_id - 1u;
 
             const auto bank_start = std::chrono::steady_clock::now();
 
@@ -1419,7 +1418,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
 
     log_info(tt::LogTest, "Persistent partitioned-workers all-DRAM test running on {} chip(s)", devices_.size());
 
-    const uint32_t chips_to_test = get_dram_max_chips_from_env(static_cast<uint32_t>(devices_.size()));
+    const uint32_t chips_to_test = get_dram_max_chips_from_env(devices_.size());
 
     const bool parallel_chips = true;
 
@@ -1480,7 +1479,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
         std::vector<std::vector<size_t>> workers_for_bank(num_dram_channels);
 
         for (size_t worker_idx = 0; worker_idx < worker_cores.size(); worker_idx++) {
-            const uint32_t bank_id = static_cast<uint32_t>(worker_idx % num_dram_channels);
+            const uint32_t bank_id = worker_idx % num_dram_channels;
             workers_for_bank[bank_id].push_back(worker_idx);
         }
 
@@ -1520,7 +1519,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
                         TT_FATAL(!bank_workers.empty(), "No workers assigned to DRAM bank {}", bank_id);
 
                         const uint64_t bytes_per_core_base =
-                            (static_cast<uint64_t>(total_bytes_per_controller) / bank_workers.size()) & ~0xFFFULL;
+                            (total_bytes_per_controller / bank_workers.size()) & ~0xFFFULL;
 
                         TT_FATAL(
                             bytes_per_core_base >= chunk_bytes,
@@ -1533,8 +1532,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
                             "bytes_per_core_base must fit uint32_t");
 
                         const uint64_t covered_bytes = bytes_per_core_base * bank_workers.size();
-                        const uint64_t remainder_bytes =
-                            static_cast<uint64_t>(total_bytes_per_controller) - covered_bytes;
+                        const uint64_t remainder_bytes = total_bytes_per_controller - covered_bytes;
 
                         TT_FATAL((remainder_bytes & 0xFFFULL) == 0ULL, "remainder_bytes must stay 4KB aligned");
 
@@ -1563,9 +1561,9 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
 
                             job.job_id = job_id++;
                             job.bank_id = bank_id;
-                            job.bank_offset_lo = static_cast<uint32_t>(bank_offset & 0xFFFFFFFFull);
-                            job.bank_offset_hi = static_cast<uint32_t>((bank_offset >> 32) & 0xFFFFFFFFull);
-                            job.total_bytes = static_cast<uint32_t>(bytes_this_core);
+                            job.bank_offset_lo = bank_offset & 0xFFFFFFFFull;
+                            job.bank_offset_hi = (bank_offset >> 32) & 0xFFFFFFFFull;
+                            job.total_bytes = bytes_this_core;
                             job.chunk_bytes = chunk_bytes;
                             job.pattern_id = pattern_id;
                             job.seed = seed;
@@ -1589,7 +1587,7 @@ TEST_F(MeshDispatchFixture, DramDeployment_PersistentPartitionedWorkersAllDramBa
             seed += advance_seed;
         }
 
-        const uint64_t total_jobs_for_chip = static_cast<uint64_t>(job_id - 1u);
+        const uint64_t total_jobs_for_chip = job_id - 1u;
 
         const auto start = std::chrono::steady_clock::now();
